Convert several loop bounds in atm_srk3 from integer pointers to simple integers
In preparation for porting of computational code that resides directly in the
atm_srk3 routine to GPUs using OpenACC directives, this commit converts the
nEdges, nEdgesSolve, and nVertLevels dimensions that are used as loop bounds
from integer pointers to simple integers. Pointers to these dimensions are
retrieved as *_ptr versions of the dimensions, and those pointers are then
assigned to integer dimensions.