Make global_list an allocatable target in mpas_block_decomp_cells_for_proc
This is a workaround for problems encountered on systems using OpenMPI
v5.x (observed with 5.0.7).  The inlist argument to
mpas_dmpar_scatter_ints seems to be affected by the MPI_ScatterV call,
which then causes run-time fails in mpas_block_decomp_cells_for_proc
when re-reading the block decomposition file. Making global_list an
allocatable, associating a pointer with it, and passing that pointer to
mpas_dmpar_scatter_ints seems to resolve the issue.