Add 3-d radar reflectivity (refl10cm) to the 'da_state' output stream
This commit adds changes to the microphysics driver to capture the 3-d 10 cm
radar reflectivity to a new field, refl10cm. The refl10cm field is only
allocated when the Thompson, Thompson aerosol-aware, or WSM6 microphysics
schemes are used.

Additionally, this commit adds the refl10cm field to the 'da_state' output
stream.